Fix compatibility issues with ncurses calls on Linux.
This commit looks to resolve a subtle incompatibility with the curses
update_panels() call on Linux. Fortunately, in lieu of this, we make
multiple calls to wnoutrefresh(), which queues screen changes to what
curses terms the virtual screen, and that are all applied to the
physical screen in one rendering with a subsequent call to doupdate().
All changes are under ifdefs so only apply to Linux systems. This has
been tested on Ubuntu running Linux kernel 5.11.