Initial implementation of the diff command.
This is still a work-in-progress; however, the basic infrastructure is
laid and the 'fnc diff' command is now functional. Most, if not all,
documented command line options and key bindings are also operational.
Preliminary tests have only been conducted on OpenBSD and macOS,
however, so further testing is required before hitting trunk.