Complete implementation of the diff command.
This commit finishes the diff command, and introduces a significant
change to the entire diff logic. We remove the bespoke verbose_diff()
routine, and instead utilise the commit manifest F card ordering to
identify new additions and removals to and from the repository, as well
as changes between versions. This removes a lot of code, and simplifies
the programming logic. Tested on OpenBSD 6.9, macOS 10.15, & Linux 5.11.