Add key bindings to jump to start and end of view.
This commit introduces support for Vim-like 'gg' and 'G' key bindings to
navigate to the start and end of the view, respectively. Conventional
HOME and END keys are also aliased to this movement. In timeline view,
the first (most recent) and last (earliest) commits in the repository
are selected. While in diff view, the screen is scrolled to the first
and last lines of the diff file. Both operations are quite performant.
The manual page still needs to be updated.