Overhaul diff implementation when diffing local changes.
This commit makes a significant change to the diff logic when local
changes on disk are being diffed. Separate routines are now called when
diffing two file artifacts in memory or when diffing a local file on
disk. This results in several improvements that produce a more reliable
diff, particularly when removed files in the current checkout are
involved.