Change 'fnc diff' behaviour when diffing local checkout.
Prior to this commit, 'fnc diff <sym>', where <sym> is any previous version
or one of Fossil's special tags, with changes on disk in the local checkout,
would _not_ show said changes. Rather, only the differences between <sym> and
the checked-out version would be displayed. This commit changes that behaviour
so that fnc is now consistent with fossil(1) whereby local changes are also
displayed when only one version is supplied to the diff command.