Update man page and add key bindings to in-app help.
Add '?' to the set of key bindings that open the in-app help. While here, add
<SPACE> to the set of key bindings that scroll down a page in the help screen.
This emulates man(1) (and less(1)) behaviour; however, we also use <SPACE> to
select a commit from the timeline so this arguably introduces inconsistency.
Nonetheless, I think it's a worthwhile addition. Also, remove unused colour
scheme initialisation in the help implementation.