Backout [47770b7de8], which broke diff view scrolling.
These changes to libfossil's fsl_appendf() break the ability to scroll in diff
view. Further testing is needed to identify which change(s) broke scrolling
through diff output so that we can either refactor our diff code to work with
the changes or see if the API can still produce the old behaviour.