Initial implementation of the new tree command.
This commit introduces 'fnc tree' and its associated documentation; it is
fully-featured except for the blame integration. The command displays a tree of
the repository and enables opening a timeline view of all commits that modified
the file(s) identified by the selected node. This adds an optional 'path' arg
to 'fnc timeline' that, similarly, displays the timeline of commits involving
the specified path. Tested on OpenBSD 6.9, macOS 10.15, and Ubuntu 21.04 with
clang and gcc-11.