Add optional path argument to the 'fnc tree' command.
This commit enables users to specify a path in the repository so that 'fnc tree'
displays a tree of all nodes at the specified path. This comports with the
optional timeline argument of the same name, introduced in [b50dcf30e0], that
displays all commits modifying the supplied path. Documentation has also been
updated to reflect this addition.