Enhance [d7cea451c8] to handle 'fnc timeline -R path' case.
If users specifies a path with no open checkout, do not process the path or
quit yet; instead, check the path against the F cards of the latest check-in,
or the specified commit. Only then, if the path is not valid, exit gracefully.
While here, remove the modeline in favour of [d0825b684f|EditorConfig] and clean
up some code to comply with our [https://man.openbsd.org/style.9|style(9)].