New threaded blame implementation and upstream libf.
This commit introduces a significant refactor of the new blame command by
annotating the file in a second thread, which greatly improves responsiveness
of the UI and generally realises better performance. The change also makes our
blame code functionally similar to our timeline code. This rewrite was made
simple thanks to the upstream change to the annotate API made by stephan@.