Don't set line->annotated for every callback step.
The previous setting was too liberal, and would fail with '--limit|-n' calls.
Only set this bit when we have a commit hash for a given line. With the
'--reverse|-r' option, we can default to the root check-in; however, we have
no blanket cover for descending chronological annotations when depth traversal
is limited, so display '..........' instead.