Enable browsing the tree of a commit from the timeline.
Add new 't' key binding to the timeline view, which opens a tree for the
currently selected commit. The tree will represent the state of the repository
as at the selected commit. In addition, if the timeline is of a given path
(i.e., fnc timeline path), the tree will jump straight to path.