Rollback handling of 'fnc invalid_command_or_path'.
In [a5fbe1193c], handling of 'fnc invalid_command_or_path' was changed to
default to 'fnc timeline'; however, this could be confusing if the user
mistyped a second argv that was intended to be a command or path. Return to
the previous handling that only defaults to 'fnc timeline' when either no
arguments are supplied, or a second argv is successfully resolved to a valid
repository path.