Add blame key bindings to the F1|?|H in-app help text.
While here, deduplicate some help text. Rather than copy the same description
strings for the UTF8 and ASCII key bindings, reuse the same description string,
and append it to the terminal-codeset-dependent set of key bindings. This makes
it easier to update key binding help text as more views or bindings are added.