Naive fix to display trees with files no longer on disk.
This is a temporary fix that needs further work; however, it addresses the
immediate problem. When selecting commits from the timeline with the 't' key
binding, or requesting a tree of a commit with 'fnc tree -c <sym>', if the
repository at the time of the requested commit had files that are no longer on
disk, the lstat() call fails. This commit uses an extremely naive heuristic
to apply an appropriate mode to the missing file path.