Display message for tree request of non-checkin artifact.
Rather than error out, display a message to the user when the 't' key binding
is used in the timeline view to request a tree of non-checkin artifacts such
as tags, technotes, or wikis. I don't particularly like the sleep() call, but
without it, the message is lost to the screen being redrawn.