Output message when file is not in the requested version.
The 'p' blame key binding requests the version of the file from the parent
commit of the currently selected line to be blamed. If the file history,
however, ends at the current commit, the file doesn't exist in its parent.
Rather than silently ignore the request as we currently do, inform the user
that the file is not in the requested version. While here, remove a couple
gotos and instead cleanup at the point of error.