Allow special casing repos with no 'initial empty check-in'
Some fossil(1) repositories, such as the canonical fnc, don't have the standard
'initial empty check-in' due to carrying a local patch to elide this commit.
This change considers such a case, and continues the diff despite having no
parent to diff against, thereby showing the entire diff of newly added files.
Prior to this commit, only the changeset was listed. This shouldn't cause any
problems, but keep an eye out for any breakage.