Assign NULL_DEVICE late to ensure diff index is populated.
In some cases where the plus(+) side of the diff is a deleted or renamed file
that is no longer on disk, the index will display "/dev/null" because we assign
this label to zplus too early. This commit corrects this mislabelling by leaving
zplus as NULL, so that the index assignment comes from the minus(-) side.