Fix timeline -> diff view bug that displays bogus changes.
Commit [f4fc5d8229] introduced a bug in the diff view of the first commit
shown in the timeline by displaying changes from the local checkout. This
happens when a check is made to determine which diff routine to call, based
on the rid of the selected commit. A change to the comparator in this check
due to changing the rid assignment of the current checkout brought about the
bug, which this commit now fixes.