Enhance 'fnc diff' to allow diff of arbitrary file blobs.
When two file artifact hash IDs are passed to 'fnc diff', display the changes
between the two blobs as stored in the database with the corresponding hashes.
While here, optimise parsing of diff types by replacing some strcmp() checks
with enums. I'm not real happy with the overloading of fnc_commit_artifact to
represent blobs, but this would be a much larger diff to refactor the logic to
handle this use case, and we can iterate on this change.