Update documentation with new 'fnc branch' command.
Both the man page and in-app help are now up-to-date with the new branch command
key bindings and options. While here, relocate more of the key bindings common
to all views to the global section, and make SPACE act as ENTER in the branch
view to open the timeline of the currently selected branch.