Major simplification of the build system by sdk@
This change removes the autosetup/configure script for the build, distilling
the procees to a make-only build; it keeps gmake, however. Tested and confirmed
to build on OpenBSD (x86 and sparc64), FreeBSD, macOS, and Linux. Pushing to
a branch for further testing on arm and odroid, but otherwise destined for
trunk pending further tests. Change sans debug bits authored by sdk@.