Remove debug make target and implicitly search for ncursesw.
For simplicity, use 'CFLAGS="-DDEBUG -g" make', and rather than search and then
include, add the atypical ncursesw include path by default; it'll be searched
and used if there—otherwise have no effect. Thus we combine the search and
inclusion in one place with the rest of the includes. Changes proposed by sdk@.