Make branch view accessible from all other views.
Add the 'b'ranch key binding to the timeline, diff, and tree views, and the
'T'ag key binding to the blame view, to open and populate a branch view with
all repository branches. Update the man page and in-app help to document this
addition. While here, reorder some documented diff key bindings for consistency.
Note: the blame view 'b' and 'B' key bindings were unavailable, hence the 'T'.
I'm not satisfied with this discrepancy, however, and alternatives are wanted.