Provide 'fnc config' interface to set configurable settings.
As per /chat discussion with stephan@ and sdr@, rather than save environment
variable values to the new fx_fnc repository table, provide an interface for
setting values in either the local or global fossil(1) configuration database.
Also, change the order of precedence so that settings are first sought locally,
then globally, and lastly from environment variables, falling back to default
fnc values if none are found. This commit implements all but the code to store
settings in the global config. The man page has been updated to document this
change. While an improvement, this adds a fair bit of code, which, although
trivial, is arguably more costly than the benefit of per-project settings.