Implement global config for the sake of comparison.
As per the suggestion from stephan@ in /chat, this commit introduces support for
global settings. While not yet convinced it's needed as envvars achieve the same
result, we can now experiment with both and make a decision as to whether
this, or local-only settings, or neither should land on trunk. In the case we
decide to keep global settings, I'm not sure that the current implementation is
right: we open and close the global config db in main so it's open for the
duration of the session. Perhaps it's better to only open it when 'fnc config'
is run, and during initial startup of the other commands when we need to detect
any global settings, but then close it for the remainder of the session.