Fix bug that broke coloured views with compiler optimisations.
With optimisations >O0, which make(1) sets as a system default in sys.mk, the
the compiler exposed a bug in the colour logic that resulted in monochrome views
irrespective of whether the user had defined colours or not. This  introduces a
fix for this behaviour so that all optimisation levels work as expected.