Revert globally configurable settings.
Due to the way all opened databases are attached to the main handle, we cannot
detach the global config db while other databases are attached with active
statements. This requires keeping the global config db open for the duration of
the session, which results in locking errors when fossil(1) commands are
executed while fnc is open. We cannot workaround this by initialising all
colour schemes at startup without also intialising all of fnc's views, which
is hugely inefficient, but would also require refactoring a lot of code. We
also cannot workaround this by finalising or resetting all statements prior to
reading the global db, because we build commits on demand, which requires an
active statement until all commits in the repository history are loaded, and it
would be suboptimal to queue all commits at startup. Given envvars provide the
exact same feature as global config, it is saner to remove the latter instead.