Implement new '--filter' timeline option and key binding.
This commit introduces support for filtering the timeline by commits with any
of the comment, user, or branch fields matching the entered pattern. When passed
on the command line, the initial timeline view will be populated with matching
commits. When the 'F' key binding is used, a new, child timeline view will be
created and populated with matching commits. Documentation has been updated with
this new feature. Still undecided on whether we should match on all these fields
as we have the --user and --branch options, so perhaps we should only match on
comments.