Make all commands compatible with the -R|--repo option.
This commit expands the --repo option to the branch and blame commands so that
all commands now work without a checkout or with a checkout different from that
covering the working directory. When this option is used, Fossil Special Tags
'current', 'prev', and 'next' are invalid arguments and operands to some of the
commands. The help and man page have been updated to document its usage.