Expand unveil to cover the 'fnc config' command.
After further experimenting and testing, it seems unveil(2) requires the
repository directory—not just the database file—and the (c)reate permission for
'fnc config'. From the [https://man.openbsd.org/unveil.2|man page] and an
understanding of what the config command does (write the SQLite db), I'd expect
that the repo db and (w)rite permission would suffice, but this results in a
FSL_RC_DB error. This commit limits unveiling the repo db directory and create
permission strictly to the config command; however, I'll conduct further
investigation to understand why the directory and create permission is needed.