Don't return error when blaming empty file and fix segv bug.
Previously, we would error out when attempting to load a blame view of an empty
file. This change simply displays the empty file with no annotation (as there
is nothing to annotate). However, a future change might display a hyperlinked
commit hash that, when selected, opens the diff view of the commit in which the
file was introduced or last modified. While here, fix UB when dereferencing a
null pointer (i.e., testing non-existent file mode of the ".." tree entry), and
remove an unnecessary strdup() allocation.