Implement configurable horizontal split view mode.
This commit introduces support for opening child views in a horizontal split.
Three modes are supported: auto, horizontal, and vertical. The default mode is
auto, which will open vertical splits if the terminal is >= 120 columns. Less
than this, and child views will open in a horizontal split of a configurable
size that defaults to 60% of the terminal height. The man page has been updated
to document this feature. Tested on OpenBSD, macOS, and Linux; however,
substantial modifications to the view logic were needed to implement this
change, so be on the lookout for breakage. Currently, only timeline to diff
child views are supported, but more combinations should soon follow.