More robust cycling of views with horizontal splits.
When cycling back to the timeline in fullscreen after scrolling the timeline
from a diff view in a horizontal split, make sure to request n commits from
the timeline thread, where n is the number of commits scrolled, to ensure the
timeline view has no empty lines. To reproduce, from the timeline, open a
commit diff view in a horizontal split, scroll down the timeline with '.'
through several commits beneath the split, toggle fullscreen ('f'), then cycle
('tab') back to the timeline.