Improve paging and don't scroll beyond final commit.
Keep the first and last displayed commit from the current screen in view as
the last and first displayed commit of the following screen when paging up and
down, respectively. This comports with paging behaviour in Vim, browsers, etc.
Also, ensure we don't scroll beyond the final commit (i.e., initial commit to
the repository) when paging the timeline in a horizontal split.