Expand horizontal split support to tree -> blame view.
This commit enables opening the blame view of a file selected from a tree view
in a horizontal split. Similar to the timeline -> diff view, split size can be
configured with the FNC_VIEW_SPLIT_HEIGHT option, which is documented in the
man page. Likewise, split behaviour is determined by the FNC_VIEW_SPLIT_MODE.