Plug small memleak when traversing commits from diff view.
When traversing the timeline from the diff view, ensure s->id{1,2} are freed
before reassigning the new selected commit hash id. When scrolling the timeline
with '.' or ',' while in diff view, we don't call close_diff_view() to dispose
of related allocations, which leaks the id string with each reassignment. This
change plugs the leak.