Improve blame handling of binary files.
When selecting a binary file from a tree view to blame, rather than error out,
display a message to screen. Similarly, display an informative error message
when a binary file is passed to the 'fnc blame' command from the command line.
While here, ensure we don't scroll beyond the final tree entry when in a
horizontal split.