Implement 'l' & 'L' key bindings to goto line in blame view.
Enable navigating to a given line in blame view with the (L)ine key binding,
which will centre the entered line in the middle of the window. If the line
number is less than 50% of the view line height, simply highlight and select
the line without redrawing the screen. Also, enable toggling the display of
file line numbers with the (l)ine key binding. Update the man page and in-app
help with this feature.  While here, replace some bitmask flag bitwise ops with
the relevant macros, and lexicographically order key bindings in the man page.