Reset diff state to clear highlights, and improve view focus.
When changing commits from the diff view with '<' or '>' with a search term
highlighted, the highlight persists in the same place on the new view. This
also occurs when loading parent or previous commit blame views, so reset state
here too. While here, improve the attempt to keep the current diff lines in
view as per [6da4ddc0241f0067].