Backout conversion to libf diff v2 API to fix large diffs.
When attempting to view the very large diff of commit [e18c2922cf47b58e] using
the new diff v2 implementation, fnc will hang on the 'generating diff' stage,
while the previous diff API processes this diff in no time. Unsure as to where
the bug is, but rolling back this change till it can be resolved.