Implement (l)ine number key binding for diff view.
Enable toggling a column of line numbers for the diff view with the 'l' key
binding. While here, improve the blame view line number display, and optimise
reset_diff_view() to better keep the current lines in view when changing context
lines, or using other key bindings (e.g., whitespace, verbose) that manipulate
the diff and change the content. Update the docs with this new feature.