Implement horizontal scroll in diff and blame views.
This commit introduces support for sideways scrolling in diff and blame views,
which required changing an existing key binding; namely, (l), which previously
toggled line numbers is now used to scroll right. All new keymaps introduced
with this commit:

  (#) now toggles line numbers
  (l) repurposed to scroll right
  (h) scroll left
  ($) scroll to end of line
  (0) scroll to beginning of line

While the horizontal scroll was trivial, rewiring of string parsing in our grep
implementation was tricky to continue working properly with the new scroll. It
has been tested and is heavily documented with comments, but lookout for any
change. While here, tune the logic that attempts to keep displayed lines in view
when using key bindings that manipulate diff content. Docs have been updated.