Implement C-{n,p} and 'F' diff view key maps for navigation.
Navigate to the next and previous files in the diff with C-n and C-p,
respectively. Similarly, the 'F' key map enables navigating directly to a given
file by opening a prompt to enter the file number. While here, rather than pass
a bag of variables to key diff routines, pass the diff state struct. And fix
a bug in the diff view file position percentage indicator.