Implement Vim-like C-{d,u} keymaps in all views.
Page up and down half a screen with C-u and C-d, respectively, in timeline,
diff, tree, blame, and branch views. This behaves like our C-{b,f,y,e} keymaps
whereby text movement is inversely related to the direction in which the view
scrolls, albeit on the scale of half a screen; that is, the view scrolls down
(or up) while the text moves up (or down). Documentation has been updated.