Enhance search behaviour in all views and dedup related code.
Prior to this commit, when grepping a pattern the view would always scroll so
that the first displayed line contains the matching term. This isn't favourable
in diff and blame views as we lose the preceding context. With this change, the
view will /not/ scroll if the next match is on the current screen, but will
instead move the selection cursor to the matching line and highlight the token.
When the next match is off-screen, the view will scroll so that the matching
line is approximately one third from the top of the view. This displays all
surrounding context. In tree and branch views, the behaviour is the same for
when matches are in view, but when matches are off-screen we scroll so that
the match occupies the first line. Preceding context is not important in
branch and tree views. While here, consolidate diff and blame grep code into
one routine.