Fix vertical split regression and DB locking issue.
[8017fee270ef46] introduced a view initialisation bug when FNC_VIEW_SPLIT_MODE
is set to "vertical" in terminal sessions that are less than 120 columns wide.
The view mode was inferred from the start column, which is 0 in such cases, and
would incorrectly set VIEW_SPLIT_HRZN, producing an error when instantiating
view dimensions. While here, _finally_ (optimistically) fix the DB lock problem.