Implement -p|--prototype option for the diff command.
This commit introduces support for displaying in the chunk header the enclosing
function that the change is in. This option is on by default in Git and CVS,
and is documented in diff(1) as --show-c-function; however, the heuristic used
in this implementation has been tested and shown to work with many C-like
languages including, but not limited to, Java, Python, JavaScript, and Rust. I
propose it should produce reliable results with most (all?) C-like source code,
but will produce irrelevant results with Lisps and markup languages such as
HTML and markdown. Documentation has been updated with this feature.