Implement side-by-side formatted diff and refactor diff renderer.
This commit introduces the --sbs CLI option and 'S' diff view key map to render
side-by-side formatted diffs. New user-definable option FNC_COLOUR_DIFF_SBS_EDIT
configures the colour for "edited" lines, which defaults to red, and the column
width is dynamically-sized per file to be wide enough to fit the longest line
in the diff. The backend is in place to enable user-defined widths, but until
we add independent sideways scroll, such an option would be useless; and I'm
undecided on the latter. A significant, albeit more expensive, refactor of the
diff renderer was needed to enable finer control over how diffs are displayed.
The cost of which I'm uncertain is worth the benefit as I'll never use SBS
diffs, although some users might. Further, this diff touches a _lot_ of code,
some that is integral to all view rendering, which I'm not too excited about.
By necessity, some of the changes are not nearly as readable as the existing
code, which is why I'm branching this for the time being. If it proves stable
after broader use, it should land on trunk. All related documentation has also
been updated. Potential road map of improvements to the SBS diff:

  1. two-tone colour highlight in edited lines like fossil ui
  2. independent sideways scroll (uncertain)
  3. full-width index header (uncertain)